% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dens.R
\name{calc_dens}
\alias{calc_dens}
\title{calculate density for each samples based on dimension reduction embedding}
\usage{
calc_dens(df_list, dens, k, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{dens}{method used to estimate density, options are GMM (Gaussian mixture model)
and KNN (K-nearest Neighbor)}

\item{k}{number of nearest neighbor, used when dens = "KNN".}

\item{x}{a data frame with row corresponding to single cell, contains the info of
each cell}

\item{sample_id}{column names in x that contains the sample ID}

\item{dim_redu_data}{a data matrix or frame contains the dimension reduction embedding
for each cell}
}
\value{
A list of length number of samples, contains the estimated density for each
sample
}
\description{
This function loads a vector of the sample names, with length of the
 number of samples whose density is to be estimated, a data frame x contains the info
 of cells, with row number equals the cell number, a data matrix or frame dim_redu
 contains dimension reduction embedding for each cell, which is used to calculate the
 density for each sample.
}
\examples{
data(example_data)
set.seed(1)
BiocParallel::register(BiocParallel::SerialParam())
sample_name = as.character(unique(example_data[, "donor_label"]))

#working with large data set, use BiocParallel
mod_list = calc_dens(sample_id = "donor_label",
                     x = example_data, dim_redu = "PC", dens = "GMM",
                     BPPARAM=BiocParallel::SerialParam())


}
