% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R
\name{calc_kl}
\alias{calc_kl}
\title{calculate  KL for a pair of samples}
\usage{
calc_kl(mclust_mod1, mclust_mod2, n = 10000, ep = 1e-64)
}
\arguments{
\item{mclust_mod1}{mclust object for sample1, contains estimated density \hat{p_1}}

\item{mclust_mod2}{mclust object for sample2, contains estimated density \hat{p_2}}

\item{n}{number of monte-carlo simulations to generate for \hat{p}}

\item{ep}{error term added to the KL divergence calculation}
}
\value{
a numeric value of distance between sample1 and sample2's distribution.
}
\description{
This function loads the metadata, which contains the sample id, disease group, dimension reduction embedding which has names
in the form "dim_i". Dimension reduction embedding is used to calculate the density for each sample, denoted by their sample id.
}
