% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{cluster_distance}
\alias{cluster_distance}
\title{Calculate KL divergence between all sample pairs' cell type proportion}
\usage{
cluster_distance(cell_sample_ids, celltype)
}
\arguments{
\item{cell_sample_ids}{a list of the samples IDs each cell comes from. Length
must match the number of element in `celltype`}

\item{celltype}{a vector of use defined cell type}
}
\value{
clusprop_dist a symmetric matrix of divergences
}
\description{
This function calculates a matrix of pairwise divergences
  between input samples' cell type proportion.
}
\examples{
# Bring in small example data of single cell embeddings
data(example_SCE_small)
sample_id <- SingleCellExperiment::colData(example_SCE_small)$sample_id 
cluster_id <- SingleCellExperiment::colData(example_SCE_small)$cluster_id 
dist_result <- cluster_distance(sample_id, cluster_id)
dist_result
}
