% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat}
\alias{distMat}
\title{calculate symmetrised  KL for each pair of samples}
\usage{
distMat(x, sample_id, dim_redu, ndim)
}
\arguments{
\item{x}{a data frame with row corresponding to single cell, contains a column of sample ID, and columns of dimension
reduction embedding}

\item{sample_id}{column names in x that contains the sample ID}

\item{dim_reduc}{dimension reduction index in x file (e.g. dim_reduc = "PC" for "PC_1").}
}
\value{
A distance matrix contains the symmetrised KL divergence value calculated for each pair of samples.
}
\description{
This function loads the metadata, which contains the sample id, disease group, dimension reduction embedding which has names
in the form "dim_i". Dimension reduction embedding is used to calculate the density for each sample, denoted by their sample id.
}
\examples{
data(example_data)
dist_mat <- SKL(example_data, "donor_label", "PC", 1:10)

#print out the distance matrix using PCA embedding.
dist_mat

}
