% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.sample_mclust}
\alias{.sample_mclust}
\title{generate monte-carlo simulation from GMM estimated density for integrating expectation for KL divergence}
\usage{
.sample_mclust(mclust_mod, n = 10000)
}
\arguments{
\item{mclust_mod}{mclust object which contains the estimated density \hat{p}}

\item{n}{number of simulations to generate. Default = 10,000.}
}
\value{
A matrix contains the symmetrised KL divergence value calculated for each pair of samples.
}
\description{
This function uses a object from mclust package as input, which contains the estimated density \hat{p} and
performs monte-carlo simulation based on \hat{p}
}
