% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{gloscope}
\alias{gloscope}
\title{calculate statistical divergence for each pair of samples}
\usage{
gloscope(
  embedding_matrix,
  cell_sample_ids,
  dens = "GMM",
  dist_mat = "KL",
  r = 10000,
  num_components = c(1:9),
  k = 50,
  BPPARAM = BiocParallel::SerialParam(),
  fit_density = NULL,
  return_density = FALSE
)
}
\arguments{
\item{embedding_matrix:}{a matrix of latent embeddings with rows corresponding to cells and columns to dimensions}

\item{cell_sample_ids:}{a list of the samples IDs each cell comes from. Length must match the number of rows in `embedding_matrix`}

\item{dens:}{the density estimation. One of c("GMM","KNN")}

\item{dist_mat:}{distance metric to calculate the distance. One of c("KL","JS")}

\item{r:}{number of monte-carlo simulations to generate}

\item{num_components:}{a vector of integers for the number of components to fit GMMS to, default is 1:9}

\item{k:}{number of k nearest negibhour for KNN density estimation, default k = 50.}

\item{BPPARAM:}{BiocParallel parameters, default is running in serial}

\item{prefit_density:}{a named list of `densityMclust` objects for each sample, default is NULL}

\item{return_density:}{return the GMM parameter list or not (if applicable), default is FALSE}
}
\value{
A matrix containing the paitwise divergence or distance between all pairs of samples
}
\description{
This function calculates a matrisx of pairwise divergences between input GloScope representations.
}
\examples{
\dontrun{
data(example_data)
sample_ids <- seurat_object[[]]$sample
pca_embeddings <- seurat_object@reductions$pca@cell.embeddings
pca_embeddings_subset <- pca_embeddings[,1:10]
dist_result <- gloscope(pca_embeddings_subset, sample_ids,
                    BPPARAM = BiocParallel::SerialParam(RNGseed=2))
dist_result
}

}
