% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gloscope_proportion.R
\name{gloscope_proportion}
\alias{gloscope_proportion}
\title{Calculate divergence between all sample pairs' cell type proportion}
\usage{
gloscope_proportion(
  cell_sample_ids,
  cell_type_ids,
  ep = 0,
  dist_metric = c("KL", "JS")
)
}
\arguments{
\item{cell_sample_ids}{a vector of the samples IDs each cell comes from. Length
must match the number of element in `cell_type_ids`}

\item{cell_type_ids}{a vector of use defined cell type}

\item{ep}{an integer of error term added to 0 proportion. Default ep = 0.}

\item{dist_metric}{distance metric to calculate the distance. One of
c("KL","JS")}
}
\value{
clusprop_dist a symmetric matrix of divergences
}
\description{
This function calculates a matrix of pairwise divergences
  between input samples' cell type proportion.
}
\examples{
# Bring in small example data of single cell embeddings
data(example_SCE_small)
sample_id <- SingleCellExperiment::colData(example_SCE_small)$sample_id 
cluster_id <- SingleCellExperiment::colData(example_SCE_small)$cluster_id 
dist_result <- gloscope_proportion(sample_id, cluster_id, ep = 0.5, 
                                   dist_metric = "KL")
dist_result
}
