% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{bigPalette}
\alias{bigPalette}
\title{prepare for plotting colors}
\format{
An object of class \code{character} of length 56.
}
\usage{
bigPalette
}
\description{
prepare for plotting colors
}
\keyword{datasets}
