% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{setBPParam}
\alias{setBPParam}
\title{Helper function to set parallelization parameters}
\usage{
setBPParam(BPPARAM, requested_cores)
}
\arguments{
\item{BPPARAM}{(list) Request BiocParallel parameters}

\item{requested_cores}{(integer) Number of cores to parallelize over if possible}
}
\value{
BPPARAM (list) BiocParallel parameters that match system availability
}
\description{
Helper function to set parallelization parameters
}
