% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dens.R
\name{calc_dens}
\alias{calc_dens}
\title{calculate density for each samples based on dimension reduction embedding}
\usage{
calc_dens(
  sample_name,
  sample_id,
  x,
  dim_redu_data,
  dens = c("GMM", "KNN"),
  k = NULL
)
}
\arguments{
\item{sample_id}{column names in x that contains the sample ID}

\item{x}{a data frame with row corresponding to single cell, contains the info of
each cell}

\item{dim_redu_data}{a data matrix or frame contains the dimension reduction embedding
for each cell}

\item{dens}{method used to estimate density, options are GMM (Gaussian mixture model)
and KNN (K-nearest Neighbor)}

\item{k}{number of nearest neighbor, used when dens = "KNN".}
}
\value{
A list of length number of samples, contains the estimated density for each
sample
}
\description{
This function loads a vector of the sample names, with length of the
 number of samples whose density is to be estimated, a data frame x contains the info
 of cells, with row number equals the cell number, a data matrix or frame dim_redu
 contains dimension reduction embedding for each cell, which is used to calculate the
 density for each sample.
}
\examples{
data(example_data)
set.seed(1)
sample_name = as.character(unique(example_data[, "donor_label"]))
dim_redu_data = example_data[,str_detect(colnames(example_data), "PC")]
mod_list = calc_dens(sample_name, sample_id = "donor_label",
                     x = example_data, dim_redu_data, dens = "GMM")


}
