% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dist.R, R/distMat.R
\name{calc_kl}
\alias{calc_kl}
\alias{calc_EMD}
\alias{calc_JS}
\alias{calc_dist}
\alias{distMat}
\title{calculate  KL for a pair of samples}
\usage{
calc_kl(
  mod_list,
  sample1,
  sample2,
  df_list,
  n,
  dens,
  k,
  ndim,
  varapp,
  epapp,
  ep
)

calc_EMD(mod_list, sample1, sample2, dens, ndim)

calc_JS(mod_list, sample1, sample2, df_list, n, dens, k, ep, ndim)

calc_dist(
  mod_list,
  s1,
  s2,
  df_list,
  n,
  dens,
  k,
  ep,
  ndim,
  dist_mat,
  varapp,
  epapp
)

distMat(
  x,
  sample_id,
  dim_redu,
  ndim,
  k = 50,
  dens = c("GMM", "KNN"),
  n = 10000,
  ep = 1e-64,
  dist_mat = c("KL", "EMD", "JS"),
  BPPARAM = BiocParallel::bpparam(),
  varapp = FALSE,
  returndens = FALSE,
  epapp = FALSE
)
}
\arguments{
\item{mod_list}{a list contains each samples' estimated density}

\item{sample1}{sample 1 index}

\item{sample2}{sample 2 index}

\item{df_list}{a list contain each samples' dimension reduction embedding}

\item{n}{number of monte-carlo simulations to generate}

\item{dens}{type of density to estimate for.}

\item{k}{number of k nearest negibhour for KNN density estimation, default k = 50.}

\item{ndim}{number of dimension reduction to keep}

\item{varapp}{logic variable for using variational approximation or not}

\item{epapp}{whether to apply the error term}

\item{ep}{error term added to the KL divergence calculation}

\item{s1}{sample 1 name}

\item{s2}{sample 2 name}

\item{dist_mat}{distance metric to calculate the distance}

\item{x}{a data frame with row corresponding to single cell, contains a column of sample ID, and columns of dimension
reduction embedding}

\item{sample_id}{column names in x that contains the sample ID}

\item{dim_redu}{dimension reduction index in x file (e.g. dim_reduc = "PC" for "PC_1").}

\item{BPPARAM}{BiocParallel parameters}

\item{returndens}{return the GMM parameter list or not}
}
\value{
a numeric value of distance between sample1 and sample2's distribution.

a numeric value of distance between sample1 and sample2's distribution.

a numeric value of distance between sample1 and sample2's distribution.

a numeric value of estimated symmatrised KL divergence between
sample1 and sample2's distribution.

A distance matrix contains the symmetrised KL divergence value calculated for each pair of samples.
}
\description{
This function loads the metadata, which contains the sample id, disease group, dimension reduction embedding which has names
in the form "dim_i". Dimension reduction embedding is used to calculate the density for each sample, denoted by their sample id.

This function loads the metadata, which contains the sample id, disease group, dimension reduction embedding which has names
in the form "dim_i". Dimension reduction embedding is used to calculate the density for each sample, denoted by their sample id.

This function loads the metadata, which contains the sample id, disease group, dimension reduction embedding which has names
in the form "dim_i". Dimension reduction embedding is used to calculate the density for each sample, denoted by their sample id.
}
\examples{
data("example_data")
set.seed(1)
dist_mat <- distMat(example_data, sample_id = "donor_label", dim_redu = "PC",
                    ndim = 10, dens = "GMM", n=10000, ep = 1e-64, dist_mat = "KL",
                    BPPARAM = BiocParallel::SerialParam(), varapp = FALSE,
                    returndens = FALSE, epapp = FALSE)

#print out the distance matrix using PCA embedding.
dist_mat

}
