% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distMat.R
\name{distMat}
\alias{distMat}
\title{calculate statistical divergence for each pair of samples}
\usage{
distMat(
  x,
  sample_id,
  dim_redu,
  ndim,
  k = 50,
  dens = "GMM",
  r = 10000,
  ep = 1e-64,
  dist_mat = "KL",
  num_components = c(1:9),
  BPPARAM = BiocParallel::SerialParam(),
  varapp = FALSE,
  returndens = FALSE,
  epapp = FALSE,
  fit_density = NULL,
  min_cell = 500,
  is_scvi = FALSE
)
}
\arguments{
\item{x}{a data frame with row corresponding to single cell, contains a column of sample ID, and columns of dimension
reduction embedding}

\item{sample_id}{column names in x that contains the sample ID}

\item{dim_redu}{dimension reduction index in x file (e.g. dim_reduc = "PC" for "PC_1").}

\item{ndim}{number of dimension reduction to keep}

\item{k}{number of k nearest negibhour for KNN density estimation, default k = 50.}

\item{dens}{type of density to estimate for.}

\item{r}{number of monte-carlo simulations to generate}

\item{ep}{error term added to the KL divergence calculation}

\item{dist_mat}{distance metric to calculate the distance}

\item{num_components}{a vector of integers for the number of components to fit GMMS to, default is 1:9}

\item{BPPARAM}{BiocParallel parameters; NULL to let system pick}

\item{varapp}{logic variable for using variational approximation or not}

\item{returndens}{return the GMM parameter list or not}

\item{epapp}{whether to apply the error term}

\item{fit_density}{a named list containing the fit density for each patient}

\item{min_cell}{the user defined minimum cell number per sample to check to have reliable results}

\item{is_scvi}{whether the dimension reduction is based on ScVI and the entire data frame would be used}
}
\value{
A distance matrix contains the symmetrised KL divergence value calculated for each pair of samples.
}
\description{
This function loads the metadata, which contains the sample id, disease group, dimension reduction embedding which has names
in the form "dim_i". Dimension reduction embedding is used to calculate the density for each sample, denoted by their sample id.
}
\examples{
data(example_data)
dist_result <- distMat(example_data, sample_id = "patient_id", dim_redu = "PC",
                    ndim = 10, dens = "KNN", r=10000, ep = 1e-64, dist_mat = "KL",
                    BPPARAM = BiocParallel::SerialParam(), varapp = FALSE,
                    returndens = FALSE, epapp = FALSE)

#print out the distance matrix using PCA embedding.
dist_result

}
