% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{example of data frame as the input for distMat}
\format{
A data frame with 5000 rows and 52 variables
}
\usage{
example_data
}
\description{
A data frame contains a subset of data
from COVID PBMC data. The data includes the patient
ID (patient_id), condition of the patients (Status),
and PCA dimenstion reduction (PC_1 to PC_50)
}
\keyword{datasets}
