% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R
\name{plotMDS}
\alias{plotMDS}
\title{plot MDS representation of the distance matrix}
\usage{
plotMDS(dist_mat, n = 10, x, sample_id, group_id)
}
\arguments{
\item{dist_mat}{the distance matrix of comparing each pairs
of patients' distributions}

\item{n}{number of MDS dimension to generate}

\item{x}{a data frame contains each patients' info}

\item{sample_id}{column name in x that contains the sample ID}

\item{group_id}{column name in x that contains the patient condition}
}
\value{
a list contains the MDS embedding and plot of the distance matrix, color-coded by patient conditions
}
\description{
This function loads the distance matrix,
and the info for each patients, including patient id
and patient condition group. It will create the MDS plot
for the distance matrix and color-coded each patient's dot
based on the condition.
}
\examples{
data("example_data")
set.seed(1)
dist_mat <- distMat(example_data, sample_id = "patient_id", dim_redu = "PC", dist_mat = "KL",
                    ndim = 10, dens = "KNN", n=10000, ep = 1e-64,
                    BPPARAM = BiocParallel::SerialParam(), varapp = FALSE,
                    returndens = FALSE, epapp = FALSE)
pat_info <- unique(example_data[, c("patient_id", "Status")])

mds_result = plotMDS(dist_mat = dist_mat, n = 2,
 x =  pat_info, "patient_id", "Status")

mds_result$plot
mds_result$mds
}
