% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMDS.R, R/utils.R
\name{plotMDS}
\alias{plotMDS}
\alias{.paletteBig}
\title{Plot the multidimensional scaling of the GloScope represenation}
\usage{
plotMDS(dist_mat, metadata_df, sample_id, group_id, n = 10)

.paletteBig()
}
\arguments{
\item{dist_mat}{The divergence matrix output of `gloscope()`}

\item{metadata_df}{A data frame contains each sample's metadata. Note this is NOT at the cell-level.}

\item{sample_id}{The column name or index in metadata_df that contains the sample ID}

\item{group_id}{The column name or index in metadata_df that contains the patient condition}

\item{n}{Number of MDS dimension to generate, default = 10}
}
\value{
A list containing the MDS embedding and plot of the distance matrix
}
\description{
This function creates a multidimensional scaling plot for
 a set of samples using their GloScope divergence. Each sample's scatter will
be color-coded based on their phenotype.
}
\examples{
\dontrun{
data(example_data)
sample_ids <- example_data$metadata$sample_id
pca_embeddings <- example_data$pca_embeddings
pca_embeddings_subset <- pca_embeddings[,1:10] # select the first 10 PCs
dist_result <- gloscope(pca_embeddings_subset, sample_ids,
                    BPPARAM = BiocParallel::SerialParam(RNGseed=2))
mds_result <- plotMDS(dist_mat = dist_result, metadata_df =  unique(example_data$metadata),
"sample_id", "phenotype",n=2)
mds_result$plot
mds_result$mds
}
}
